Fix saving image to gallery.

+++ app/src/main/java/fi/tamk/tuni/paintapp/MainActivity.java
@@ -1,11 +1,15 @@
 package fi.tamk.tuni.paintapp;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.provider.MediaStore;
+import android.util.Log;
 import android.view.View;
 import android.view.WindowManager;
 import android.widget.ImageButton;
@@ -18,6 +22,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     private PaintView mPaintView;
     private ImageButton mCurrentPaint, mBrushButton, mEraseButton, mNewButton, mSaveButton, mUndoButton, mRedoButton;
     private float mXtraSmallBrush, mSmallBrush, mMediumBrush, mLargeBrush;
+    private static final int PERMISSION_REQUEST_CODE = 100;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -141,23 +146,27 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 newDialog.show();
                 break;
             case R.id.button_save:
-                AlertDialog.Builder saveDialog = new AlertDialog.Builder(this);
-                saveDialog.setTitle("Save drawing");
-                saveDialog.setMessage("Save drawing to device Gallery?");
-                saveDialog.setPositiveButton("Yes", (l, w) -> {
-                    mPaintView.setDrawingCacheEnabled(true);
-                    String imgSaved = MediaStore.Images.Media.insertImage(
-                            getContentResolver(), mPaintView.getDrawingCache(),
-                            UUID.randomUUID().toString() + ".png", "drawing");
-                    if (imgSaved != null) {
-                        Toast.makeText(getApplicationContext(), "Drawing saved to Gallery!", Toast.LENGTH_SHORT).show();
-                        mPaintView.destroyDrawingCache();
-                    } else {
-                        Toast.makeText(getApplicationContext(), "Oops! Image could not be saved.", Toast.LENGTH_SHORT).show();
-                    }
-                });
-                saveDialog.setNegativeButton("Cancel", (l, w) -> l.cancel());
-                saveDialog.show();
+                if (checkPermission()) {
+                    AlertDialog.Builder saveDialog = new AlertDialog.Builder(this);
+                    saveDialog.setTitle("Save drawing");
+                    saveDialog.setMessage("Save drawing to device Gallery?");
+                    saveDialog.setPositiveButton("Yes", (l, w) -> {
+                        mPaintView.setDrawingCacheEnabled(true);
+                        String imgSaved = MediaStore.Images.Media.insertImage(
+                                getContentResolver(), mPaintView.getDrawingCache(),
+                                UUID.randomUUID().toString() + ".png", "drawing");
+                        if (imgSaved != null) {
+                            Toast.makeText(getApplicationContext(), "Drawing saved to Gallery!", Toast.LENGTH_SHORT).show();
+                            mPaintView.destroyDrawingCache();
+                        } else {
+                            Toast.makeText(getApplicationContext(), "Oops! Image could not be saved.", Toast.LENGTH_SHORT).show();
+                        }
+                    });
+                    saveDialog.setNegativeButton("Cancel", (l, w) -> l.cancel());
+                    saveDialog.show();
+                } else {
+                    requestPermission();
+                }
                 break;
             case R.id.button_undo:
                 // Undo button logic
@@ -167,6 +176,32 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 // Redo button logic
                 mPaintView.redoRecentPath();
                 break;
+        }
+    }
+
+    private void requestPermission() {
+        if (ActivityCompat.shouldShowRequestPermissionRationale(MainActivity.this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+            Toast.makeText(MainActivity.this, "Write External Storage permission allows us to save files. Please allow this permission in App Settings.", Toast.LENGTH_LONG).show();
+        } else {
+            ActivityCompat.requestPermissions(MainActivity.this, new String[]{android.Manifest.permission.WRITE_EXTERNAL_STORAGE}, PERMISSION_REQUEST_CODE);
+        }
+    }
+
+    private boolean checkPermission() {
+        int result = ContextCompat.checkSelfPermission(MainActivity.this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
+        return result == PackageManager.PERMISSION_GRANTED;
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        switch (requestCode) {
+            case PERMISSION_REQUEST_CODE:
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                Log.e("value", "Permission Granted, Now you can use local drive .");
+            } else {
+                Log.e("value", "Permission Denied, You cannot use local drive .");
+            }
+            break;
         }
     }
 }
